requirements.txt
torch
torchaudio
transformers
peft
accelerate
datasets==2.19.0
librosa
jiwer
soundfile
evaluate
pandas
tensorboard
ipywidgets
pyannote.audio
pyannote-whisper
